Included the open source tool EMMA into the build process to generate a code coverage report when ever the unit tests are run. The report can be found in target/emma/report.
I have been using EMMA for a while at work and am very happy with the
results. It is not quite as polished as Clover, but you get what you pay
for. I think it is good to be able to get a code coverage report
whenever the tests are run.